+++ of_to_tp.py
@@ -19,10 +19,6 @@ from omnifocus import build_model, find_database
 import os
 import codecs
 
-'''
-This is a visitor that dumps out an OPML file containing each and every entry
-in the database. You can also specify a particular type, e.g. just 'Task'.
-'''
 class PrintTaskpaperVisitor(Visitor):
     def __init__ (self, out, depth=0):
         self.depth = depth
@@ -33,13 +29,16 @@ class PrintTaskpaperVisitor(Visitor):
     def end_folder (self, folder):
         self.depth-=1
     def begin_project (self, project):
-        print >>self.out, self.tabs() + project.name + ':'
+        if self.project_mode:
+            print >>self.out, self.tabs() + project.name + ':'
+        else:
+            print >>self.out, self.tabs() + '- ' + self.remove_trailing_colon(project.name) + self.tags(project)
         self.depth+=1
     def end_project (self, project):
         self.depth-=1
     def begin_task (self, task):
-        if len(task.children) == 0:
-            print >>self.out, self.tabs() + '- ' + task.name + self.tags(task)
+        if self.is_empty (task) or self.project_mode == False:
+            print >>self.out, self.tabs() + '- ' + self.remove_trailing_colon(task.name) + self.tags(task)
         else:
             print >>self.out, self.tabs() + task.name + ':'
         self.depth+=1
@@ -67,6 +66,12 @@ class PrintTaskpaperVisitor(Visitor):
         if len (tags) > 0:
             return ' ' + ' '.join(tags)
         return ''
+    def remove_trailing_colon (self, string):
+        if string.endswith(':'):
+            return string[:-1]
+        return string
+    def is_empty (self, item):
+        return len ([x for x in item.children if x.marked]) == 0
     def tabs (self):
         return '\t' * (self.depth)
 
@@ -79,9 +84,9 @@ if __name__ == "__main__":
     out=codecs.open(file_name, 'w', 'utf-8')
     
     print >>out, 'Projects:'
-    traverse_list (PrintTaskpaperVisitor (out, depth=1), root_projects_and_folders)
+    traverse_list (PrintTaskpaperVisitor (out, True, depth=1), root_projects_and_folders)
     print >>out, 'Contexts:'
-    traverse_list (PrintTaskpaperVisitor (out, depth=1), root_contexts)
+    traverse_list (PrintTaskpaperVisitor (out, False, depth=1), root_contexts, project_mode=False)
     
     out.close()