+++ visitors.py
@@ -14,16 +14,17 @@ See the License for the specific language governing permissions and
 limitations under the License.
 '''
 
-from treemodel import Visitor, TASK, PROJECT
+from treemodel import Visitor, TASK, PROJECT, CONTEXT
 from datematch import process_date_specifier
 import re
 from datetime import datetime
 
 INCLUDED='INCLUDED'
+EXCLUDED='EXCLUDED'
 PATH_TO_INCLUDED='PATH_TO_INCLUDED'
 
 def match_name (item, regexp):
-        return re.search (regexp, item.name) != None
+    return re.search (regexp, item.name) != None
 
 def match_date_against_range (thedate, date_range):
     start, end = date_range
@@ -50,46 +51,72 @@ def match_completed (item, date_range):
 def match_flagged (item, ignore):
         return item.flagged
 
-def set_attrib_to_root (item, name, value):
-    if item != None:
+def set_attrib_to_root (path_to_root, name, value):
+    for item in path_to_root:
         item.attribs[name] = value
-        set_attrib_to_root (item.parent, name, value)
-    
+
+def mark_branch_not_marked (item, project_mode):
+    if item.marked:
+        item.marked = False
+        if (item.type == TASK or item.type == PROJECT) and not project_mode:
+            # We only got here because we recursed from a context
+            # Tasks/Projects are not a tree in context mode, they're flat so we don't want
+            # to un-mark all the children since they might be in a different context
+            return
+        for child in item.children:
+            mark_branch_not_marked (child, project_mode)
+            
 class BaseFilterVisitor(Visitor):
     def __init__(self, include=True):
         self.filter = None
         self.include = include
+        self.traversal_path = []
     def begin_any (self, item):
-        
+        # Can't use the item parent since this only has meaning
+        # in project mode - have to track our own traversal path
+        parent = None
+        if len(self.traversal_path) > 0:
+            parent = self.traversal_path[-1]
         item.attribs[PATH_TO_INCLUDED] = False
-        if item.parent != None:
-            # Inherit this attribute
-            item.attribs[INCLUDED] = item.parent.attribs[INCLUDED]
+        if parent != None:
+            # Inherit these attribute
+            item.attribs[INCLUDED] = parent.attribs[INCLUDED]
+            item.attribs[EXCLUDED] = parent.attribs[EXCLUDED]
+            assert parent.attribs[INCLUDED] != None, "missing attribute in " + parent.name
+            assert parent.attribs[EXCLUDED] != None, "missing attribute in " + parent.name
         else:
             item.attribs[INCLUDED] = False
+            item.attribs[EXCLUDED] = False
+        self.traversal_path.append(item)
     def end_any (self, item):
+        assert item.attribs[PATH_TO_INCLUDED] != None, "missing attribute in " + item.name
+        assert item.attribs[INCLUDED] != None, "missing attribute in " + item.name
+        assert item.attribs[EXCLUDED] != None, "missing attribute in " + item.name
+        self.traversal_path.pop()
         if self.include and not (item.attribs[INCLUDED] or item.attribs[PATH_TO_INCLUDED]):
-            item.marked = False
+            mark_branch_not_marked (item, self.project_mode)
         # We've finished processing the node, tidy up
         # and avoid confusing the next filter.
         del (item.attribs[INCLUDED])
+        del (item.attribs[EXCLUDED])
         del (item.attribs[PATH_TO_INCLUDED])
+    def match_required (self, item):
+        if item.attribs[INCLUDED] or item.attribs[EXCLUDED]:
+            # The decision has already been made
+            return False
+        return True
     def set_item_matched (self, item, matched):
         # invoked from begin_XXX
-        if item.attribs[INCLUDED]:
-            # The decision has already been made
-            return
-    
         if self.include:
             if matched:
                 # Then we want this node in the output and want to stop
                 # this filter testing removing any parents or children of this node
                 item.attribs[INCLUDED] = True
-                set_attrib_to_root (item.parent, PATH_TO_INCLUDED, True)
+                set_attrib_to_root (self.traversal_path, PATH_TO_INCLUDED, True)
         else: # In exclude mode
             if matched:
                 # This node is toast
-                item.marked = False
+                mark_branch_not_marked (item, self.project_mode)
             else:
                 # We haven't excluded it so it stays
                 pass
@@ -101,8 +128,9 @@ class AnyNameFilterVisitor(BaseFilterVisitor):
         self.match_fn = match_name
     def begin_any (self, item):
         BaseFilterVisitor.begin_any (self, item)
-        matched = self.match_fn(item, self.filter)
-        self.set_item_matched(item, matched);
+        if self.match_required(item):
+            matched = self.match_fn(item, self.filter)
+            self.set_item_matched(item, matched);
     def __str__(self):
         return 'name ' + includes (self.include) + ' "' + self.filter + '"'
     
@@ -112,9 +140,10 @@ class AnyFlaggedFilterVisitor(BaseFilterVisitor):
         self.match_fn = match_flagged
     def begin_any (self, item):
         BaseFilterVisitor.begin_any (self, item)
-        if item.type == PROJECT or item.type == TASK:
-            matched = self.match_fn(item, self.filter)
-            self.set_item_matched(item, matched);
+        if self.match_required(item):
+            if item.type == PROJECT or item.type == TASK:
+                matched = self.match_fn(item, self.filter)
+                self.set_item_matched(item, matched);
     def __str__(self):
         return 'name ' + includes (self.include) + ' flagged'
 
@@ -124,8 +153,9 @@ class FolderFilterVisitor(BaseFilterVisitor):
         self.filter = filtr
         self.match_fn = match_fn
     def begin_folder (self, folder):
-        matched = self.match_fn(folder, self.filter)
-        self.set_item_matched(folder, matched);
+        if self.match_required(folder):
+            matched = self.match_fn(folder, self.filter)
+            self.set_item_matched(folder, matched);
             
 class TaskFilterVisitor(BaseFilterVisitor):
     def __init__(self, filtr, match_fn, include=True):
@@ -133,8 +163,9 @@ class TaskFilterVisitor(BaseFilterVisitor):
         self.filter = filtr
         self.match_fn = match_fn
     def begin_task (self, task):
-        matched = self.match_fn(task, self.filter)
-        self.set_item_matched(task, matched);
+        if self.match_required(task):
+            matched = self.match_fn(task, self.filter)
+            self.set_item_matched(task, matched);
             
 class ProjectFilterVisitor(BaseFilterVisitor):
     def __init__(self, filtr, match_fn, include=True):
@@ -142,8 +173,9 @@ class ProjectFilterVisitor(BaseFilterVisitor):
         self.filter = filtr
         self.match_fn = match_fn
     def begin_project (self, project):
-        matched = self.match_fn(project, self.filter)
-        self.set_item_matched(project, matched);
+        if self.match_required(project):
+            matched = self.match_fn(project, self.filter)
+            self.set_item_matched(project, matched);
             
 class ContextFilterVisitor(BaseFilterVisitor):
     def __init__(self, filtr, match_fn, include=True):
@@ -151,8 +183,9 @@ class ContextFilterVisitor(BaseFilterVisitor):
         self.filter = filtr
         self.match_fn = match_fn
     def begin_context (self, context):
-        matched = self.match_fn(context, self.filter)
-        self.set_item_matched(context, matched);
+        if self.match_required(context):
+            matched = self.match_fn(context, self.filter)
+            self.set_item_matched(context, matched);
 
 def includes (include):
     if include:
@@ -247,22 +280,6 @@ class TaskFlaggedFilterVisitor(TaskFilterVisitor):
     def __str__(self):
         return 'Task flagged ' + includes (self.include) + ' flagged'
 
-class FlatteningVisitor (Visitor):
-    def __init__(self):
-        self.projects = []
-    def begin_project (self, project):
-        self.projects.append(project)
-    def end_task (self, task):
-        mypos = task.parent.children.index (task)
-        # Add this nodes children above itself,
-        # this is the omnifocus way.
-        for child in task.children:
-            task.parent.children.insert (mypos, child)
-            child.parent = task.parent
-        task.children = []
-    def __str__ (self):
-        return 'Flatten'
-        
 class TaskCompletionSortingVisitor (Visitor):
     def end_project (self, project):
         project.children.sort(key=lambda item:self.get_key(item))
@@ -278,10 +295,48 @@ class FolderNameSortingVisitor (Visitor):
         folder.children.sort(key=lambda item:item.name)
     def __str__ (self):
         return 'Folders/Projects sorted by name'
-    
+
+def flatten (item):
+    new_children = []
+    for child in item.children:
+        # Add this nodes children above itself,
+        # this is the omnifocus way.
+        flatten (child)
+        new_children = new_children + child.children
+        new_children.append(child)
+        child.children = []
+    item.children = new_children
+
+class FlatteningVisitor (Visitor):
+    def __init__(self):
+        self.projects = []
+        self.contexts = []
+    def end_project (self, project):
+        self.projects.append(project)
+        flatten (project)
+        for child in project.children:
+            child.parent = project        
+    def end_task (self, task):
+        flatten (task)
+        for child in task.children:
+            child.parent = task
+    def end_context (self, context):
+        new_children = []
+        for child in context.children:
+            if child.type != CONTEXT:
+                new_children.append (child)
+                child.parent = None
+        context.children = new_children
+        self.contexts.append(context)   
+    def __str__ (self):
+        return 'Flatten'
+
 class PruningFilterVisitor (Visitor):
     def end_project (self, project):
-        self.prune_if_empty(project)
+        if self.project_mode:
+            self.prune_if_empty(project)
+        else:
+            project.match = False
     def end_folder (self, folder):
         self.prune_if_empty(folder)
     def end_context (self, context):