+++ CHANGELOG.md
@@ -1,5 +1,13 @@
 # Changes #
 
+## 1.0.4 (2013-04-15) ##
+
+- Added Context mode.
+- Improved test script.
+- Filter performance improvement.
+- Improved flattening algorithm to work in Context mode.
+- Added "No Context" context that all tasks without a context get added to.
+
 ## 1.0.3 (2013-04-13) ##
 
 - Big reworking of filter logic to squash a design bug.